enhydrator/src/test/java/com/airhacks/enhydrator/out/NamedSinkTest.java
package com.airhacks.enhydrator.out;

/*
 * #%L
 * enhydrator
 * %%
 * Copyright (C) 2014 Adam Bien
 * %%
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *      http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * #L%
 */

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;
import org.junit.Test;

/**
 *
 * @author airhacks.com
 */
public class NamedSinkTest {

    @Test
    public void defaultName() {
        NamedSink sink = new CustomSink();
        String name = sink.getName();
        assertThat(name, is("customSink"));
    }

    @Test
    public void shortName() {
        NamedSink sink = new Z();
        String name = sink.getName();
        assertThat(name, is("z"));

    }

}
